+++ packages/terra-form-checkbox/tests/wdio/form-checkbox-field-spec.js
@@ -27,33 +27,35 @@ Terra.describeViewports('CheckboxField', ['tiny', 'large'], () => {
 Terra.describeViewports('CheckboxField', ['tiny'], () => {
   it('should display CheckboxField interactions valid state', () => {
     browser.url('/raw/tests/terra-form-checkbox/form-checkbox/checkbox-field/controlled-checkbox-field');
-    browser.waitForVisible('#testing-checkbox-field');
-    browser.click('[for="website-dept"]');
-    browser.click('[for="ux-dept"]');
+
+    $('#testing-checkbox-field').waitForDisplayed();
+    $('[for="website-dept"]').click();
+    $('[for="ux-dept"]').click();
 
     Terra.validates.element('interactions with valid state');
   });
 
   it('should display CheckboxField interactions invalid state', () => {
-    browser.click('[for="website-dept"]');
-    browser.click('[for="ux-dept"]');
+    $('[for="website-dept"]').click();
+    $('[for="ux-dept"]').click();
 
     Terra.validates.element('interactions with invalid state');
   });
 
   it('should display CheckboxField interactions valid State with hideRequired', () => {
     browser.url('/raw/tests/terra-form-checkbox/form-checkbox/checkbox-field/hide-required-invalid-checkbox-field');
-    browser.waitForVisible('#testing-checkbox-field');
-    browser.click('[for="website-dept"]');
-    browser.click('[for="ux-dept"]');
+
+    $('#testing-checkbox-field').waitForDisplayed();
+    $('[for="website-dept"]').click();
+    $('[for="ux-dept"]').click();
 
     Terra.validates.element('interactions valid State with hideRequired');
   });
 
   it('should display CheckboxField interactions invalid state with hideRequired', () => {
-    browser.waitForVisible('#testing-checkbox-field');
-    browser.click('[for="website-dept"]');
-    browser.click('[for="ux-dept"]');
+    $('#testing-checkbox-field').waitForDisplayed();
+    $('[for="website-dept"]').click();
+    $('[for="ux-dept"]').click();
 
     Terra.validates.element('interactions invalid state with hideRequired');
   });