+++ packages/terra-arrange/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Upgrade tests to terra-functional-testing and WebDriverIO v6
+  * Removed unnecessary `wdio.conf.js` file
+
 ## 3.43.0 - (February 2, 2021)
 
 * Changed
@@ -11,7 +15,7 @@
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 3.41.0 - (December 8, 2020)
 
 * Changed