+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Removed unnecessary `wdio.conf.js` file
+  * Updated to terra-functional-testing
+
 ## 3.50.0 - (March 2, 2021)
 
 * Changed