+++ packages/terra-paginator/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Removed unnecessary `wdio.conf.js` file
+  * Upgrade tests to terra-functional-testing and WebDriverIO v6
+
 ## 2.70.0 - (March 2, 2021)
 
 * Changed
@@ -16,7 +20,7 @@
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 2.67.0 - (December 8, 2020)
 
 * Changed